feat: add conditional notifications with threshold context for speed tests
- Show specific notification messages for threshold violations (high ping, low download/upload)
- Display actual threshold values in notification messages when available
- Add distinct message prefixes: [!] for alerts, [SPEEDTEST] for completion
- Enhance agent and packet loss notifications to show thresholds
- Rename database files: notifications_types.go for clarity
- Remove legacy notification configuration from config.toml
